requirements.txt
_libgcc_mutex=0.1=main
_pytorch_select=0.1=cpu_0
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py38h32f6830_1
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=py_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h1e0a361_1000
c-ares=1.16.1=h516909a_0
ca-certificates=2020.6.24=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py38_0
cffi=1.14.0=py38he30daa8_1
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=pyh9f0ad1d_0
cryptography=3.0=py38h766eaa4_0
cudatoolkit=10.1.243=h6bb024c_0
filelock=3.0.12=pyh9f0ad1d_0
flake8=3.8.3=py_0
freetype=2.10.2=h5ab3b9f_0
gast=0.3.3=py_0
google-auth=1.19.2=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyh8c360ce_0
gperftools=2.7=h767d802_2
grpcio=1.30.0=py38h2c89da0_0
h5py=2.10.0=nompi_py38hfb01d0b_104
hdf5=1.10.6=nompi_h3c11f04_100
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py38h32f6830_0
intel-openmp=2020.1=217
joblib=0.16.0=py_0
jpeg=9b=h024ee3a_2
kaggle=1.5.6=py38_1
keras-preprocessing=1.1.0=py_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.12.3=h8b12597_2
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py_0
mccabe=0.6.1=py38_1
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
mock=4.0.2=py_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py38hfd86e86_0
nltk=3.5=py_0
numpy=1.18.5=py38ha1c710e_0
numpy-base=1.18.5=py38hde5b4d6_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
openssl=1.1.1g=h7b6447c_0
opt_einsum=3.3.0=py_0
packaging=20.4=pyh9f0ad1d_0
perl=5.26.2=h516909a_1006
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.2=py38_0
protobuf=3.12.3=py38h950e882_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=pyh9f0ad1d_2
pyflakes=2.2.0=py_0
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.3=hcff3b4d_2
python-dateutil=2.8.1=py_0
python-slugify=4.0.1=pyh9f0ad1d_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.1.243_cudnn7.6.3_0
readline=8.0=h7b6447c_0
regex=2020.7.14=py38h1e0a361_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
sacremoses=0.0.43=pyh9f0ad1d_0
scipy=1.5.0=py38h0b6359f_0
sentencepiece=0.1.91=py38hbf85e49_1
setuptools=49.2.0=py38_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
tensorflow=2.2.0=mkl_py38h6d3daf0_0
tensorflow-base=2.2.0=mkl_py38h5059a2d_0
tensorflow-estimator=2.2.0=pyh95af2a2_0
termcolor=1.1.0=py_2
text-unidecode=1.3=py_0
tk=8.6.10=hbc83047_0
torchvision=0.7.0=py38_cu101
tqdm=4.48.0=pyh9f0ad1d_0
transformers=3.0.2=py_0
unidecode=1.1.1=py_0
urllib3=1.24.3=py38_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py38_0
wrapt=1.12.1=py38h1e0a361_1
xz=5.2.5=h7b6447c_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
fast-bleu==0.0.86
future==0.18.2
rouge-score==0.0.4
scikit-learn==0.23.2
sentence-transformers==0.3.4
threadpoolctl==2.1.0
tokenizers==0.8.1rc1

